Backend/user_service/requirements.txt
-i https://pypi.org/simple
asgiref
django
django-cors-headers
django-mysql
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
exceptiongroup
iniconfig
packaging
pika
pluggy
psycopg2-binary
pytest
sqlparse
tomli
typing-extensions
gunicorn
channels
channels-redis[cryptography]
django-environ
daphne
pytest
pytest-django
pytest-mock
pillow
twisted[http2,tls]
